products and profiles admin views secured with @login required

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -67,8 +68,14 @@ def artwork_details(request, product_id):
     return render(request, 'products/artwork_details.html', context)
 
 
+@login_required
 def create_product(request):
+
     # Add an art product to the store
+    if not request.user.is_superuser:
+        messages.error(request, 'Unable to complete this action, please contact the site administrator')
+        return redirect(reverse('home'))
+
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -88,8 +95,14 @@ def create_product(request):
     return render(request, template, context)
 
 
+@login_required
 def update_product(request, product_id):
+
     # Update a product in the store
+    if not request.user.is_superuser:
+        messages.error(request, 'Unable to complete this action, please contact the site administrator')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -112,9 +125,13 @@ def update_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
         
     # Delete a product from the store
+    if not request.user.is_superuser:
+        messages.error(request, 'Unable to complete this action, please contact the site administrator')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
